+++ interfaces/broadcaster.go
@@ -5,11 +5,11 @@ import (
 
 	"github.com/mezonai/mmn/block"
 	"github.com/mezonai/mmn/consensus"
-	"github.com/mezonai/mmn/types"
+	"github.com/mezonai/mmn/transaction"
 )
 
 type Broadcaster interface {
 	BroadcastBlock(ctx context.Context, blk *block.Block) error
 	BroadcastVote(ctx context.Context, vt *consensus.Vote) error
-	TxBroadcast(ctx context.Context, tx *types.Transaction) error
+	TxBroadcast(ctx context.Context, tx *transaction.Transaction) error
 }
