+++ p2p/pubsub.go
@@ -3,9 +3,7 @@ package p2p
 import (
 	"context"
 	"crypto/ed25519"
-	"encoding/json"
 	"fmt"
-	"time"
 
 	"github.com/mezonai/mmn/block"
 	"github.com/mezonai/mmn/blockstore"
@@ -15,55 +13,8 @@ import (
 	"github.com/mezonai/mmn/ledger"
 	"github.com/mezonai/mmn/logx"
 	"github.com/mezonai/mmn/mempool"
-	"github.com/mezonai/mmn/types"
-
-	"github.com/libp2p/go-libp2p/core/network"
-	"github.com/libp2p/go-libp2p/core/peer"
-	ma "github.com/multiformats/go-multiaddr"
+	"github.com/mezonai/mmn/transaction"
 )
-
-func (ln *Libp2pNetwork) handleNodeInfoStream(s network.Stream) {
-	defer s.Close()
-
-	buf := make([]byte, 2048)
-	n, err := s.Read(buf)
-	if err != nil {
-		logx.Error("NETWORK:HANDLE NODE INFOR STREAM", "Failed to read from bootstrap: ", err)
-		return
-	}
-
-	var msg map[string]interface{}
-	if err := json.Unmarshal(buf[:n], &msg); err != nil {
-		logx.Error("NETWORK:HANDLE NODE INFOR STREAM", "Failed to unmarshal peer info: ", err)
-		return
-	}
-
-	newPeerIDStr := msg["new_peer_id"].(string)
-	newPeerID, err := peer.Decode(newPeerIDStr)
-	if err != nil {
-		logx.Error("NETWORK:HANDLE NODE INFOR STREAM", "Invalid peer ID: ", newPeerIDStr)
-		return
-	}
-
-	addrStrs := msg["addrs"].([]interface{})
-	var addrs []ma.Multiaddr
-	for _, a := range addrStrs {
-		maddr, err := ma.NewMultiaddr(a.(string))
-		if err == nil {
-			addrs = append(addrs, maddr)
-		}
-	}
-
-	peerInfo := peer.AddrInfo{
-		ID:    newPeerID,
-		Addrs: addrs,
-	}
-
-	err = ln.host.Connect(context.Background(), peerInfo)
-	if err != nil {
-		logx.Error("NETWORK:HANDLE NODE INFOR STREAM", "Failed to connect to new peer: ", err)
-	}
-}
 
 func (ln *Libp2pNetwork) SetupCallbacks(ld *ledger.Ledger, privKey ed25519.PrivateKey, self config.NodeConfig, bs blockstore.Store, collector *consensus.Collector, mp *mempool.Mempool) {
 	ln.SetCallbacks(Callbacks{
@@ -148,7 +99,7 @@ func (ln *Libp2pNetwork) SetupCallbacks(ld *ledger.Ledger, privKey ed25519.Priva
 
 			return nil
 		},
-		OnTransactionReceived: func(txData *types.Transaction) error {
+		OnTransactionReceived: func(txData *transaction.Transaction) error {
 			logx.Info("TX", "Processing received transaction from P2P network")
 
 			// Add transaction to mempool
@@ -275,125 +226,5 @@ func (ln *Libp2pNetwork) SetCallbacks(cbs Callbacks) {
 	}
 	if cbs.OnSyncResponseReceived != nil {
 		ln.onSyncResponseReceived = cbs.OnSyncResponseReceived
-	}
-}
-
-func (ln *Libp2pNetwork) TxBroadcast(ctx context.Context, tx *types.Transaction) error {
-	logx.Info("TX", "Broadcasting transaction to network")
-	txData, err := json.Marshal(tx)
-	if err != nil {
-		return fmt.Errorf("failed to serialize transaction: %w", err)
-	}
-
-	if err := ln.topicTxs.Publish(ctx, txData); err != nil {
-		return fmt.Errorf("failed to publish transaction: %w", err)
-	}
-
-	return nil
-}
-
-func (ln *Libp2pNetwork) BroadcastVote(ctx context.Context, vote *consensus.Vote) error {
-	msg := VoteMessage{
-		Slot:      vote.Slot,
-		BlockHash: fmt.Sprintf("%x", vote.BlockHash),
-		VoterID:   vote.VoterID,
-		Signature: vote.Signature,
-	}
-
-	data, err := json.Marshal(msg)
-	if err != nil {
-		return err
-	}
-
-	if ln.topicVotes != nil {
-		ln.topicVotes.Publish(ctx, data)
-	}
-	return nil
-}
-
-func (ln *Libp2pNetwork) BroadcastBlock(ctx context.Context, blk *block.Block) error {
-	logx.Info("BLOCK", "Broadcasting block: slot=", blk.Slot)
-
-	data, err := json.Marshal(blk)
-	if err != nil {
-		logx.Error("BLOCK", "Failed to marshal block: ", err)
-		return err
-	}
-
-	if ln.topicBlocks != nil {
-		logx.Info("BLOCK", "Publishing block to pubsub topic")
-		if err := ln.topicBlocks.Publish(ctx, data); err != nil {
-			logx.Error("BLOCK", "Failed to publish block:", err)
-		}
-	}
-	return nil
-}
-
-func (ln *Libp2pNetwork) GetPeersConnected() int {
-	return len(ln.peers)
-}
-
-func (ln *Libp2pNetwork) cleanupOldSyncRequests() {
-	ln.syncMu.Lock()
-	defer ln.syncMu.Unlock()
-
-	now := time.Now()
-	for requestID, info := range ln.activeSyncRequests {
-		if !info.IsActive || now.Sub(info.StartTime) > 5*time.Minute {
-			delete(ln.activeSyncRequests, requestID)
-		}
-	}
-}
-
-// when no peers connected the blocks will not sync must run after 30s if synced stop sync
-func (ln *Libp2pNetwork) startPeriodicSyncCheck(bs blockstore.Store) {
-	// wait network setup
-	time.Sleep(10 * time.Second)
-	ticker := time.NewTicker(30 * time.Second)
-	defer ticker.Stop()
-
-	for {
-		select {
-		case <-ticker.C:
-			ln.cleanupOldSyncRequests()
-		case <-ln.ctx.Done():
-			return
-		}
-	}
-}
-
-func (ln *Libp2pNetwork) startCleanupRoutine() {
-	ticker := time.NewTicker(1 * time.Minute)
-	defer ticker.Stop()
-
-	for {
-		select {
-		case <-ticker.C:
-			ln.CleanupExpiredRequests()
-		case <-ln.ctx.Done():
-			logx.Info("NETWORK:CLEANUP", "Stopping cleanup routine")
-			return
-		}
-	}
-}
-
-func (ln *Libp2pNetwork) startInitialSync(bs blockstore.Store) {
-	// wait network setup
-	time.Sleep(2 * time.Second)
-
-	ctx := context.Background()
-
-	if _, err := ln.RequestLatestSlotFromPeers(ctx); err != nil {
-		logx.Warn("NETWORK:SYNC BLOCK", "Failed to request latest slot from peers:", err)
-	}
-
-	var fromSlot uint64 = 0
-	localLatestSlot := bs.GetLatestSlot()
-	if localLatestSlot > 0 {
-		fromSlot = localLatestSlot + 1
-	}
-
-	if err := ln.RequestBlockSync(ctx, fromSlot); err != nil {
-		logx.Error("NETWORK:SYNC BLOCK", "Failed to send initial sync request: %v", err)
 	}
 }
