+++ p2p/transaction.go
@@ -3,9 +3,10 @@ package p2p
 import (
 	"context"
 	"encoding/json"
+	"fmt"
 
 	"github.com/mezonai/mmn/logx"
-	"github.com/mezonai/mmn/types"
+	"github.com/mezonai/mmn/transaction"
 
 	pubsub "github.com/libp2p/go-libp2p-pubsub"
 )
@@ -31,7 +32,7 @@ func (ln *Libp2pNetwork) HandleTransactionTopic(ctx context.Context, sub *pubsub
 				continue
 			}
 
-			var tx *types.Transaction
+			var tx *transaction.Transaction
 			if err := json.Unmarshal(msg.Data, &tx); err != nil {
 				logx.Warn("NETWORK:TX", "Unmarshal error:", err)
 				continue
@@ -42,4 +43,18 @@ func (ln *Libp2pNetwork) HandleTransactionTopic(ctx context.Context, sub *pubsub
 			}
 		}
 	}
+}
+
+func (ln *Libp2pNetwork) TxBroadcast(ctx context.Context, tx *transaction.Transaction) error {
+	logx.Info("TX", "Broadcasting transaction to network")
+	txData, err := json.Marshal(tx)
+	if err != nil {
+		return fmt.Errorf("failed to serialize transaction: %w", err)
+	}
+
+	if err := ln.topicTxs.Publish(ctx, txData); err != nil {
+		return fmt.Errorf("failed to publish transaction: %w", err)
+	}
+
+	return nil
 }
