+++ utils/convert.go
@@ -4,7 +4,7 @@ import (
 	"encoding/json"
 	"fmt"
 
-	"github.com/mezonai/mmn/types"
+	"github.com/mezonai/mmn/transaction"
 
 	"github.com/mezonai/mmn/block"
 	"github.com/mezonai/mmn/poh"
@@ -78,14 +78,14 @@ func ToProtoEntries(entries []poh.Entry) []*pb.Entry {
 
 // -- Tx --
 
-func ParseTx(data []byte) (*types.Transaction, error) {
-	var tx types.Transaction
+func ParseTx(data []byte) (*transaction.Transaction, error) {
+	var tx transaction.Transaction
 	err := json.Unmarshal(data, &tx)
 	return &tx, err
 }
 
-func FromProtoSignedTx(pbTx *pb.SignedTxMsg) (*types.Transaction, error) {
-	return &types.Transaction{
+func FromProtoSignedTx(pbTx *pb.SignedTxMsg) (*transaction.Transaction, error) {
+	return &transaction.Transaction{
 		Type:      pbTx.TxMsg.Type,
 		Sender:    pbTx.TxMsg.Sender,
 		Recipient: pbTx.TxMsg.Recipient,
@@ -97,14 +97,14 @@ func FromProtoSignedTx(pbTx *pb.SignedTxMsg) (*types.Transaction, error) {
 	}, nil
 }
 
-func ToProtoSignedTx(tx *types.Transaction) *pb.SignedTxMsg {
+func ToProtoSignedTx(tx *transaction.Transaction) *pb.SignedTxMsg {
 	return &pb.SignedTxMsg{
 		TxMsg:     ToProtoTx(tx),
 		Signature: tx.Signature,
 	}
 }
 
-func ToProtoTx(tx *types.Transaction) *pb.TxMsg {
+func ToProtoTx(tx *transaction.Transaction) *pb.TxMsg {
 	return &pb.TxMsg{
 		Type:      tx.Type,
 		Sender:    tx.Sender,
